+++ crates/task-impls/Cargo.toml
@@ -26,7 +26,6 @@ hotshot-task = { path = "../task" }
 hotshot-types = { path = "../types" }
 hotshot-builder-api = { path = "../builder-api" }
 jf-primitives = { workspace = true }
-memoize = { workspace = true }
 rand = { workspace = true }
 serde = { workspace = true }
 sha2 = { workspace = true }
